Move new rule back into old rule
