+++ packages/eslint-plugin/src/rules/no-bad-reference.ts
@@ -1,59 +1,163 @@
-import { TSESTree } from "@typescript-eslint/utils";
-import { commentsMatching, createRule } from "../util";
+import { ESLintUtils, TSESTree } from "@typescript-eslint/utils";
+import { createRule, findTypesPackage } from "../util";
+import * as ts from "typescript";
+import path from "path";
 import { isDeclarationPath } from "@definitelytyped/utils";
 
-type MessageId = "referencePathPackage" | "referencePathTest" | "referencePathOldVersion";
 const rule = createRule({
   name: "no-bad-reference",
   defaultOptions: [],
   meta: {
     type: "problem",
     docs: {
-      description: `Forbids <reference path="./vNN"/> in all files, <reference path="../etc"/> in declaration files, and all <reference path> in test files.`,
+      description:
+        "Forbids bad references, including those that resolve outside of the package or path references in non-declaration files.",
       recommended: "error",
     },
     messages: {
-      referencePathPackage:
-        "Don't use <reference path> to reference another package. Use an import or <reference types> instead.",
-      referencePathTest:
-        "Don't use <reference path> in test files. Use <reference types> or include the file in 'tsconfig.json'.",
-      referencePathOldVersion: "Don't use <reference path> to reference an old version of the current package.",
+      relativeImport:
+        'The import "{{text}}" resolves outside of the package. Use a bare import to reference other packages.',
+      relativeReference:
+        'The reference "{{text}}" resolves outside of the package. Use a global reference to reference other packages.',
+      testReference:
+        'The path reference "{{text}}" is disallowed outside declaration files. Use "<reference types>" or include the file in tsconfig instead.',
     },
     schema: [],
   },
   create(context) {
-    const isDeclarationFile = isDeclarationPath(context.getFilename());
-    commentsMatching(context.getSourceCode(), /<reference\s+path\s*=\s*"(.+)"\s*\/>/, (ref, comment) => {
-      if (ref.match(/^\.\/v\d+(?:\.\d+)?(?:\/.*)?$/)) {
-        report(comment, "referencePathOldVersion");
-      }
-      if (isDeclarationFile) {
-        if (ref.startsWith("..")) {
-          report(comment, "referencePathPackage");
-        }
+    const containingFileName = context.getFilename();
+    const typesPackage = findTypesPackage(containingFileName);
+    if (!typesPackage) {
+      return {};
+    }
+
+    const realNamePlusSlash = typesPackage.realName + "/";
+    function isRelativeOrSelf(name: string) {
+      return name.startsWith(".") || name.startsWith(realNamePlusSlash);
+    }
+
+    const ast = context.getSourceCode().ast;
+    const parserServices = ESLintUtils.getParserServices(context, true);
+    const sourceFile = parserServices.esTreeNodeToTSNodeMap.get(ast);
+
+    const refs: Reference[] = [];
+    for (const ref of sourceFile.referencedFiles) {
+      if (isDeclarationPath(containingFileName)) {
+        refs.push({ kind: "path", text: ref.fileName, range: ref });
       } else {
-        report(comment, "referencePathTest");
+        context.report({
+          messageId: "testReference",
+          loc: tsRangeToESLintLocation(ref, sourceFile),
+          data: { text: ref.fileName },
+        });
       }
-    });
+    }
+    for (const ref of sourceFile.typeReferenceDirectives) {
+      if (isRelativeOrSelf(ref.fileName)) {
+        refs.push({ kind: "types", text: ref.fileName, range: ref });
+      }
+    }
+    for (const ref of imports(sourceFile)) {
+      if (isRelativeOrSelf(ref.text)) {
+        refs.push({ kind: "import", text: ref.text, range: ref });
+      }
+    }
 
-    return {};
+    for (const ref of refs) {
+      const p = ref.text.startsWith(realNamePlusSlash)
+        ? path.posix.join(typesPackage.dir, ref.text.slice(realNamePlusSlash.length))
+        : ref.text;
 
-    function report(comment: TSESTree.Comment, messageId: MessageId) {
+      // TODO(jakebailey): Rather than using path.resolve, manually walk each
+      // part seeing if any of them escape the package, which would let us
+      // catch places where people relatively move out of the package and back
+      // in again.
+      //
+      // As a perf trick, we can use path.resolve if the path doesn't contain
+      // ".."; then we know that it could only ever go into a child package.
+      const resolved = path.resolve(path.dirname(containingFileName), p);
+      const otherPackage = findTypesPackage(resolved);
+
+      if (otherPackage && otherPackage.dir === typesPackage.dir) {
+        continue;
+      }
+
       context.report({
-        loc: {
-          end: {
-            column: comment.value.lastIndexOf(`"`),
-            line: comment.loc.end.line,
-          },
-          start: {
-            column: comment.value.indexOf(`"`) + 1,
-            line: comment.loc.start.line,
-          },
-        },
-        messageId,
+        messageId: ref.kind === "import" ? "relativeImport" : "relativeReference",
+        loc: tsRangeToESLintLocation(ref.range, sourceFile),
+        data: { text: ref.text },
       });
     }
+
+    return {};
   },
 });
+
+function tsRangeToESLintLocation(range: ts.TextRange, sourceFile: ts.SourceFile): TSESTree.SourceLocation {
+  const pos = sourceFile.getLineAndCharacterOfPosition(range.pos);
+  const end = sourceFile.getLineAndCharacterOfPosition(range.end);
+  return {
+    start: { line: pos.line + 1, column: pos.character + 1 },
+    end: { line: end.line + 1, column: end.character + 1 },
+  };
+}
+
+interface Reference {
+  kind: "path" | "import" | "types";
+  text: string;
+  range: ts.TextRange;
+}
+
+/**
+ * All strings referenced in `import` statements.
+ * Does *not* include <reference> directives.
+ */
+function imports({ statements }: ts.SourceFile): Iterable<ts.StringLiteralLike> {
+  const result: ts.StringLiteralLike[] = [];
+  for (const node of statements) {
+    recur(node);
+  }
+  return result;
+
+  function recur(node: ts.Node) {
+    switch (node.kind) {
+      case ts.SyntaxKind.ImportDeclaration:
+      case ts.SyntaxKind.ExportDeclaration: {
+        const { moduleSpecifier } = node as ts.ImportDeclaration | ts.ExportDeclaration;
+        if (moduleSpecifier && moduleSpecifier.kind === ts.SyntaxKind.StringLiteral) {
+          result.push(moduleSpecifier as ts.StringLiteral);
+        }
+        break;
+      }
+
+      case ts.SyntaxKind.ImportEqualsDeclaration: {
+        const { moduleReference } = node as ts.ImportEqualsDeclaration;
+        if (moduleReference.kind === ts.SyntaxKind.ExternalModuleReference) {
+          result.push(parseRequire(moduleReference));
+        }
+        break;
+      }
+
+      case ts.SyntaxKind.ImportType: {
+        const { argument } = node as ts.ImportTypeNode;
+        if (ts.isLiteralTypeNode(argument) && ts.isStringLiteral(argument.literal)) {
+          result.push(argument.literal);
+        }
+        break;
+      }
+
+      default:
+        ts.forEachChild(node, recur);
+    }
+  }
+}
+
+function parseRequire(reference: ts.ExternalModuleReference): ts.StringLiteralLike {
+  const { expression } = reference;
+  if (!expression || !ts.isStringLiteral(expression)) {
+    throw new Error(`Bad 'import =' reference: ${reference.getText()}`);
+  }
+  return expression;
+}
 
 export = rule;
