+++ packages/eslint-plugin/test/no-bad-reference.test.ts
@@ -1,7 +1,7 @@
-import * as noBadReference from "../src/rules/no-bad-reference";
 import { runTestsWithFixtures } from "./fixtureTester";
+import * as rule from "../src/rules/no-bad-reference";
 
-runTestsWithFixtures("@definitelytyped/no-bad-reference", noBadReference, {
+runTestsWithFixtures("@definitelytyped/no-bad-reference", rule, {
   valid: [
     {
       filename: "types/foo/index.d.ts",
@@ -9,49 +9,110 @@ runTestsWithFixtures("@definitelytyped/no-bad-reference", noBadReference, {
   ],
   invalid: [
     {
+      filename: "types/no-relative-references/index.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "./v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../foo" } },
+        { messageId: "relativeImport", data: { text: "./v1" } },
+        { messageId: "relativeImport", data: { text: "../foo/v1" } },
+        { messageId: "relativeImport", data: { text: "no-relative-references/v1" } },
+      ],
+    },
+    {
+      filename: "types/no-relative-references/other/other.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../foo" } },
+        { messageId: "relativeImport", data: { text: "../v1" } },
+        { messageId: "relativeImport", data: { text: "../../foo/v1" } },
+      ],
+    },
+    {
+      filename: "types/no-relative-references/v1/index.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../foo/v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../foo" } },
+        { messageId: "relativeImport", data: { text: "../../foo/v1" } },
+        { messageId: "relativeImport", data: { text: "../index" } },
+        { messageId: "relativeImport", data: { text: "../index" } },
+      ],
+    },
+    {
+      filename: "types/no-relative-references/v1/other/other.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../../foo" } },
+        { messageId: "relativeImport", data: { text: "../../../foo/v1" } },
+      ],
+    },
+    {
+      filename: "types/scoped__no-relative-references/index.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "./v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../foo" } },
+        { messageId: "relativeImport", data: { text: "./v1" } },
+        { messageId: "relativeImport", data: { text: "../foo/v1" } },
+        { messageId: "relativeImport", data: { text: "@scoped/no-relative-references/v1" } },
+      ],
+    },
+    {
+      filename: "types/scoped__no-relative-references/other/other.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../foo" } },
+        { messageId: "relativeImport", data: { text: "../v1" } },
+        { messageId: "relativeImport", data: { text: "../../foo/v1" } },
+      ],
+    },
+    {
+      filename: "types/scoped__no-relative-references/v1/index.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../foo/v1/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../foo" } },
+        { messageId: "relativeImport", data: { text: "../../foo/v1" } },
+        { messageId: "relativeImport", data: { text: "../index" } },
+        { messageId: "relativeImport", data: { text: "../index" } },
+      ],
+    },
+    {
+      filename: "types/scoped__no-relative-references/v1/other/other.d.ts",
+      errors: [
+        { messageId: "relativeReference", data: { text: "../../../foo/index.d.ts" } },
+        { messageId: "relativeReference", data: { text: "../../../foo/v1/index.d.ts" } },
+        { messageId: "relativeImport", data: { text: "../../../foo" } },
+        { messageId: "relativeImport", data: { text: "../../../foo/v1" } },
+      ],
+    },
+    {
       filename: "types/no-bad-reference/index.d.ts",
       errors: [
-        {
-          messageId: "referencePathPackage",
-          line: 1,
-          column: 20,
-          endColumn: 28,
-        },
-        {
-          messageId: "referencePathOldVersion",
-          line: 2,
-          column: 20,
-          endColumn: 25,
-        },
-        {
-          messageId: "referencePathOldVersion",
-          line: 3,
-          column: 20,
-          endColumn: 31,
-        },
-        {
-          messageId: "referencePathOldVersion",
-          line: 4,
-          column: 20,
-          endColumn: 37,
-        },
-        {
-          messageId: "referencePathOldVersion",
-          line: 5,
-          column: 20,
-          endColumn: 26,
-        },
-        {
-          messageId: "referencePathOldVersion",
-          line: 6,
-          column: 20,
-          endColumn: 32,
-        },
+        { messageId: "relativeReference", data: { text: "../other" } },
+        { messageId: "relativeReference", data: { text: "./v11" } },
+        { messageId: "relativeReference", data: { text: "./v11/index" } },
+        { messageId: "relativeReference", data: { text: "./v11/subdir/file" } },
+        { messageId: "relativeReference", data: { text: "./v0.1" } },
+        { messageId: "relativeReference", data: { text: "./v0.1/index" } },
       ],
     },
     {
       filename: "types/no-bad-reference/no-bad-reference-tests.ts",
-      errors: [{ messageId: "referencePathTest" }, { messageId: "referencePathTest" }],
+      errors: [
+        { messageId: "testReference", data: { text: "../other" } },
+        { messageId: "testReference", data: { text: "other" } },
+      ],
     },
   ],
 });
